feat: added throwable exceptions

+++ DistLedgerServer/src/main/java/pt/tecnico/distledger/server/domain/ServerState.java
@@ -1,7 +1,9 @@
 package pt.tecnico.distledger.server.domain;
 
 import pt.tecnico.distledger.server.domain.operation.*;
+import pt.tecnico.distledger.server.domain.exceptions.*;
 
+import java.io.NotActiveException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -30,11 +32,17 @@ public class ServerState {
         this.active = active;
     }
 
-    public void activate() {
+    public void activate() throws AlreadyActiveException {
+        if (this.active == true) {
+            throw new AlreadyActiveException();
+        }
         setActive(true);
     }
 
-    public void deactivate() {
+    public void deactivate() throws NotActiveException {
+        if (this.active == false) {
+            throw new NotActiveException();
+        }
         setActive(false);
     }
 
@@ -42,7 +50,10 @@ public class ServerState {
         if (debug) System.err.println(message);
     }
 
-    public int getBalance(String userId) {
+    public int getBalance(String userId) throws UserDoesNotExistException {
+        if (!this.containsUser(userId)) {
+            throw new UserDoesNotExistException();
+        }
         return accountMap.get(userId);
     }
 
@@ -58,19 +69,48 @@ public class ServerState {
         accountMap.put("broker", 1000);
     }
 
-    public void createAccount(String userId) {
+    public void createAccount(String userId) throws UserAlreadyExistsEception {
+        if (this.containsUser(userId)) {
+            throw new UserAlreadyExistsEception();
+        }
         accountMap.put(userId, 0);
         CreateOp createOp = new CreateOp(userId);
         ledger.add(createOp);
     }
 
-    public void deleteAccount(String userId) {
+    public void deleteAccount(String userId) throws BrokerCantBeDeletedException, BalanceNotZeroException, UserDoesNotExistException {
+        if (userId.equals("broker")) {
+            throw new BrokerCantBeDeletedException();
+        }
+        else if (!this.containsUser(userId)) {
+            throw new UserDoesNotExistException();
+        }
+        else if (this.getBalance(userId) == 0) {
+            throw new BalanceNotZeroException();
+        }
         accountMap.remove(userId);
         DeleteOp deleteOp = new DeleteOp(userId);
         ledger.add(deleteOp);
     }
 
-    public void transfer(String fromAccount, String toAccount, int amount) {
+    public void transfer(String fromAccount, String toAccount, int amount) throws SourceUserDoesNotExistException, 
+            DestinationUserDoesNotExistException, SourceEqualsDestinationUserException,
+                InvalidUserBalanceException, InvalidBalanceAmountException, UserDoesNotExistException {
+        if (!this.containsUser(fromAccount)) {
+            throw new SourceUserDoesNotExistException();
+        }
+        else if (!this.containsUser(toAccount)) {
+            throw new DestinationUserDoesNotExistException();
+        }
+        else if (fromAccount.equals(toAccount)) {
+            throw new SourceEqualsDestinationUserException();
+        }
+        else if (this.getBalance(fromAccount) < amount) {
+            throw new InvalidUserBalanceException();
+        }
+        else if (amount <= 0) {
+            throw new InvalidBalanceAmountException();
+        }
         accountMap.put(fromAccount, accountMap.get(fromAccount) - amount);
         accountMap.put(toAccount, accountMap.get(toAccount) + amount);
         TransferOp transferOp = new TransferOp(fromAccount, toAccount, amount);
